requirements.txt
websockets>=11.0.3
aiohttp>=3.8.5
python-telegram-bot>=20.6
python-dotenv>=1.0.0
langchain-ollama>=0.1.1
langgraph>=0.2.0
langchain-core>=0.2.0